+++ tsup.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
     devtools: "src/devtools/index.ts",
     "devtools/react": "src/devtools/react.tsx",
     react: "src/react/index.ts",
+    svelte: "src/svelte/index.ts",
     "adapters/postgres": "src/adapters/postgres/index.ts",
     "adapters/firestore": "src/adapters/firestore/index.ts",
     "storage/cloudinary": "src/storage/cloudinary/index.ts",